feat: add last modified date in bookmark

+++ tap_heap/streams.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import datetime
 import random
 import typing as t
 from pathlib import Path
@@ -21,10 +22,16 @@ class HeapTableStream(HeapStream):
     ) -> t.Iterable[dict | tuple[dict, dict | None]]:
         """Download avro files and send records."""
         self.logger.info(self.starting_replication_key_value)
+        replication_value_parts = str(self.starting_replication_key_value).split(",")
+        if len(replication_value_parts) == 2:
+            replication_sync_id, last_modified_date = replication_value_parts
+        else:
+            replication_sync_id = self.starting_replication_key_value
+            last_modified_date = None
         table_manifests, should_create_new_version = self.filter_manifests_to_sync(
             manifests,
             table_name,
-            self.starting_replication_key_value,
+            replication_sync_id,
         )
 
         sync_files = self.get_files_to_sync(
@@ -32,25 +39,39 @@ class HeapTableStream(HeapStream):
             self.config["bucket"],
         )
 
-        s3_client = s3.get_s3_client(self.config["bucket"])
+        s3_resource = s3.get_s3_resource(self.config["bucket"])
+        s3_client = s3.get_s3_client()
         for sync_id, files in sync_files.items():
-            for file in files:
-                self.logger.info(f"Syncing file : {file}")  # noqa: G004
-                line_number = 0
-                local_file_suffix = random.randint(100000,999999)  # noqa: S311
-                s3.download_file(s3_client, file, local_file_suffix)
-                with Path(f"file_{local_file_suffix}.avro").open(mode="rb") as fo:
-                    iterator = fastavro.reader(fo)
-                    for row in iterator:
-                        line_number += 1
-                        yield self.add_additional_info(row=row, sync_id=sync_id)
-                # Delete the file
-                Path(f"file_{local_file_suffix}.avro").unlink()
-            # Send state message after sync_id finish
-            self._increment_stream_state(
-                {self.replication_key: sync_id}, context=context,
-            )
-            self._write_state_message()
+            if files:
+                for file in files:
+                    file_modified_datetime = s3.get_modified_date(
+                        s3_client, self.config["bucket"], file,
+                    )
+                    if (
+                        last_modified_date is None or file_modified_datetime >= datetime.datetime.strptime(  # noqa: E501
+                            last_modified_date,
+                            r"%Y-%m-%d %H:%M:%S%z",  # ISO-8601
+                        )
+                    ):
+                        self.logger.info(f"Syncing file : {file}")  # noqa: G004
+                        line_number = 0
+                        local_file_suffix = random.randint(100000, 999999)  # noqa: S311
+                        s3.download_file(s3_resource, file, local_file_suffix)
+                        with Path(f"file_{local_file_suffix}.avro").open(mode="rb") as fo:
+                            iterator = fastavro.reader(fo)
+                            for row in iterator:
+                                line_number += 1
+                                yield self.add_additional_info(row=row, sync_id=f"{sync_id},{file_modified_datetime}")  # noqa: E501
+                        # Delete the file
+                        Path(f"file_{local_file_suffix}.avro").unlink()
+                    else:
+                        self.logger.info(f"Skip file : {file}")  # noqa: G004
+                # Send state message after sync_id finish
+                self._increment_stream_state(
+                    {self.replication_key: f"{sync_id},{file_modified_datetime}"},
+                    context=context,
+                )
+                self._write_state_message()
 
     def filter_manifests_to_sync(
         self,
@@ -65,12 +86,12 @@ class HeapTableStream(HeapStream):
         table_manifests = {
             dump_id: manifest.get(table_name)
             for dump_id, manifest in manifests.items()
-            if dump_id >= replication_key and manifest.get(table_name)
+            if dump_id >= int(replication_key) and manifest.get(table_name)
         }
 
         return (table_manifests, True)
 
-    def get_files_to_sync(self, table_manifests: dict, bucket: str)-> dict:
+    def get_files_to_sync(self, table_manifests: dict, bucket: str) -> dict:
         """Get flattened file names and remove the prefix."""
         files = {}
         for dump_id, manifest in table_manifests.items():
